requirements.txt
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1710977767585/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1704011227531/work
autobahn @ file:///home/conda/feedstock_root/build_artifacts/autobahn_1686741996912/work
Automat @ file:///home/conda/feedstock_root/build_artifacts/automat_1667331175863/work
bcrypt @ file:///D:/bld/bcrypt_1702664014679/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1698890160476/work
bokeh @ file:///home/conda/feedstock_root/build_artifacts/bokeh_1710486007020/work
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///D:/bld/cffi_1696001770413/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click @ file:///D:/bld/click_1692312014553/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
constantly==15.1.0
contourpy @ file:///D:/bld/contourpy_1699041552441/work
cryptography @ file:///D:/bld/cryptography-split_1708780364147/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
cytoolz @ file:///D:/bld/cytoolz_1706897209240/work
daphne @ file:///home/conda/feedstock_root/build_artifacts/daphne_1707583909223/work
dash @ file:///home/conda/feedstock_root/build_artifacts/dash_1710075327199/work
dash-bootstrap-components @ file:///home/conda/feedstock_root/build_artifacts/dash-bootstrap-components_1694086999250/work
dash-core-components @ file:///home/conda/feedstock_root/build_artifacts/dash-core-components_1645162191927/work
dash-html-components @ file:///home/conda/feedstock_root/build_artifacts/dash-html-components_1645162180756/work
dash-table @ file:///home/conda/feedstock_root/build_artifacts/dash-table_1645162199872/work
dash_auth @ file:///home/conda/feedstock_root/build_artifacts/dash-auth_1710820505657/work
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1710510974052/work
dask-expr @ file:///home/conda/feedstock_root/build_artifacts/dask-expr_1710761608904/work
debugpy==1.8.1
distributed @ file:///home/conda/feedstock_root/build_artifacts/distributed_1710520284602/work
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1701689956975/work
dpd_components @ file:///home/conda/feedstock_root/build_artifacts/dpd_components_1707587165039/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1707043907952/work
fonttools @ file:///D:/bld/fonttools_1710865675688/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1710808267764/work
hyperlink @ file:///home/conda/feedstock_root/build_artifacts/hyperlink_1610092164190/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
incremental @ file:///home/conda/feedstock_root/build_artifacts/incremental_1665859450441/work
ipykernel==6.29.3
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1704966972576/work
jupyter_client==8.6.1
kiwisolver @ file:///D:/bld/kiwisolver_1695380078296/work
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
lz4 @ file:///D:/bld/lz4_1704831298995/work
MarkupSafe @ file:///D:/bld/markupsafe_1706900129905/work
matplotlib @ file:///D:/bld/matplotlib-suite_1708026620064/work
msgpack @ file:///D:/bld/msgpack-python_1700926661619/work
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas @ file:///D:/bld/pandas_1708708763933/work
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1695667515973/work
pillow @ file:///D:/bld/pillow_1704252252810/work
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1708020413888/work
ply==3.11
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1700596371886/work
pyasn1 @ file:///home/conda/feedstock_root/build_artifacts/pyasn1_1701287008248/work
pyasn1-modules @ file:///home/conda/feedstock_root/build_artifacts/pyasn1-modules_1695107857548/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1706660063483/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1709721012883/work
PyQt5==5.15.9
PyQt5-sip @ file:///D:/bld/pyqt-split_1695417970429/work/pyqt_sip
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
pywin32==306
PyYAML @ file:///D:/bld/pyyaml_1695373635661/work
pyzmq==25.1.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
retrying==1.3.3
scipy @ file:///C:/bld/scipy-split_1706041618395/work/dist/scipy-1.12.0-cp311-cp311-win_amd64.whl#sha256=c70c65a82df57d892e53532fdbe573417454df6dbf355fc0ba4cec59391d26df
service-identity @ file:///home/conda/feedstock_root/build_artifacts/service-identity-build_1700936484042/work
sip @ file:///D:/bld/sip_1697300573532/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sortedcontainers @ file:///home/conda/feedstock_root/build_artifacts/sortedcontainers_1621217038088/work
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1681817562700/work
tblib @ file:///home/conda/feedstock_root/build_artifacts/tblib_1702066284995/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1692026804430/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1706112571092/work
tornado @ file:///D:/bld/tornado_1708363244199/work
Twisted @ file:///D:/bld/twisted_1709332375410/work
twisted-iocpsupport @ file:///D:/bld/twisted-iocpsupport_1695469048038/work
txaio @ file:///home/conda/feedstock_root/build_artifacts/txaio_1673804636823/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1698235201373/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
xyzservices @ file:///home/conda/feedstock_root/build_artifacts/xyzservices_1698325309404/work
zict @ file:///home/conda/feedstock_root/build_artifacts/zict_1681770155528/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
zope.interface @ file:///D:/bld/zope.interface_1708077713466/work
